Split Sentence Problem 1.0
This is my first attempt to solve the problem. I encountered an issue, and that is the last word is added to the list without its last letter.

Also, I changed the order on which I defined the 'word' variable to outside the loop. That is because previously, each time the 'start' value increases, the 'word' variable is emptied.